Add show/remove/edit

+++ main.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"strings"
 
 	"github.com/fatih/color"
 	"github.com/manifoldco/promptui"
@@ -36,15 +37,15 @@ func handleArg(db *gorm.DB) {
 	if len(os.Args) > 1 {
 		switch os.Args[1] {
 		case "-l":
-			fmt.Printf("Afficher toutes les commandes\n")
+			showCmd(db)
 		case "-h":
 			for _, a := range arrayArg {
 				color.Yellow("%v\n", a)
 			}
 		case "-a":
-			promptAdd()
+			promptAdd(db)
 		case "-e":
-			promptEdit()
+			promptEdit(db)
 		case "-r":
 			promptRemove(db)
 		}
@@ -53,7 +54,21 @@ func handleArg(db *gorm.DB) {
 	}
 }
 
-func promptAdd() {
+func showCmd(db *gorm.DB) {
+	var res []Documentation
+	db.Find(&res)
+	if len(res) > 0 {
+		for _, r := range res {
+			color.Yellow("%v :\n", r.Title)
+			color.Green("Command : %v\n", r.Command)
+			color.White("%v\n\n", r.Definition)
+		}
+	} else {
+		color.Red("No doc to show")
+	}
+}
+
+func promptAdd(db *gorm.DB) {
 	validate := func(input string) error {
 		if len(input) <= 0 {
 			return errors.New("Value too short")
@@ -69,6 +84,7 @@ func promptAdd() {
 		color.Red("Prompt failed : %v\n", err)
 		return
 	}
+
 	prompt = promptui.Prompt{
 		Label:    "Command",
 		Validate: validate,
@@ -78,6 +94,7 @@ func promptAdd() {
 		color.Red("Prompt failed :: %v\n", err)
 		return
 	}
+
 	prompt = promptui.Prompt{
 		Label:    "Description",
 		Validate: validate,
@@ -87,32 +104,60 @@ func promptAdd() {
 		color.Red("Prompt failed : %v\n", err)
 		return
 	}
-	color.Green("You choosed : %v, %v, %v", resTitleCmd, resCmd, resDescCmd)
-}
-func promptRemove(db *gorm.DB) {
-	var res []Documentation
-	db.Find(&res)
-	displayRes := make([]string, 0, len(res))
-	resDict := make(map[int]uint)
-	for i, r := range res {
-		displayRes = append(displayRes, r.Command+" - "+r.Title+" - "+r.Definition)
-		resDict[i] = r.ID
-	}
-	prompt := promptui.Select{
-		Label: "Select Day",
-		Items: displayRes,
-	}
 
-	index, result, err := prompt.Run()
-
+	prompt = promptui.Prompt{
+		Label:     "Confirm",
+		IsConfirm: true,
+	}
+	confirm, err := prompt.Run()
 	if err != nil {
 		fmt.Printf("Prompt failed %v\n", err)
 		return
 	}
-	db.Delete(&Documentation{}, resDict[index])
-	color.Green("You succesfully removed :%v %v", index, result)
+
+	if strings.ToLower(confirm) == "y" {
+		doc := Documentation{Title: resTitleCmd, Command: resCmd, Definition: resDescCmd}
+		db.Create(&doc)
+		color.Green("Created.\nYou can see it by using 'doc -l'")
+	} else {
+		color.Red("Creation canceled")
+	}
+}
+func promptRemove(db *gorm.DB) {
+	var res []Documentation
+	db.Find(&res)
+	if len(res) > 0 {
+		displayRes := make([]string, 0, len(res))
+		resDict := make(map[int]uint)
+		for i, r := range res {
+			displayRes = append(displayRes, r.Command+" - "+r.Title+" - "+r.Definition)
+			resDict[i] = r.ID
+		}
+		prompt := promptui.Select{
+			Label: "Select Day",
+			Items: displayRes,
+		}
+
+		index, result, err := prompt.Run()
+
+		if err != nil {
+			fmt.Printf("Prompt failed %v\n", err)
+			return
+		}
+		db.Delete(&Documentation{}, resDict[index])
+		color.Green("You succesfully removed :%v %v", index, result)
+	} else {
+		color.Red("No doc to remove")
+	}
+
 }
 
-func promptEdit() {
+func promptEdit(db *gorm.DB) {
+	var res []Documentation
+	db.Find(&res)
+	if len(res) > 0 {
 
+	} else {
+		color.Red("No doc to edit")
+	}
 }
